requirements.txt
Flask==3.0.3
pandas==2.2.2
flask-cors==5.0.0
pypdf==4.2.0
tabula-py==2.9.0
gunicorn==22.0.0
requests==2.31.0
python-dotenv==1.0.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
